requirements.txt
asgiref==3.5.1
beautifulsoup4==4.11.1
bokeh==2.4.3
Django==4.0.4
django-bootstrap5==21.3
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.11.3
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
numpy==1.22.4
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tornado==6.1
typing_extensions==4.2.0
zipp==3.8.0
